JenkinsFiles/ci/requirements.txt
slackclient
pathlib
PyYAML
